Feat: 디렉토리 변경 및 로그인 기능 구현
- Service/auth 디렉토리로 편입: 도메인 별 분리
- login:
  - authenticationManager를 통해 아이디/비밀번호 대조
  - jwtProvider 통해 access 및 refresh 토큰 생성
  - 두 토큰을 HttpOnly 쿠키로 저장
  - 요청으로 들어왔던 email 반환
- 반환 데이터를, 로직 성공 작동 여부를 반환할 수 있는 호출 방식으로 변환할 것
- 프론트에게 HttpOnly 쿠키 처리 및 주기적 refresh 하도록 요청할 것
